Remodeling Frontend & Backend (#1) V2.0
* Testing V2.0

* WIP: V2.0 Testing

* Delete src/Backgrounds/Particles/Particles.css

Unnecessary file due to it's migration to Tailwind.

* Delete src/SpotlightCard/SpotlightCard.css

Unnecessary file due to it's migration to Tailwind.

* Delete src/components/DiscordCTA/DiscordCTA.css

Unnecessary file due to it's migration to Tailwind.

* Delete src/components/Features/Features.css

Unnecessary file due to it's migration to Tailwind.

* Delete src/components/HeroSection/HeroSection.css

Unnecessary file due to it's migration to Tailwind.

* Delete src/components/Navbar/Navbar.css

Unnecessary file due to it's migration to Tailwind.

* Remove hover color from Discord icon

Fixing stylistic simple mistake.